+++ bricktracker/sql/set/base/base.sql
@@ -1,7 +1,9 @@
 SELECT
     {% block id %}{% endblock %}
     "bricktracker_sets"."storage",
+    "bricktracker_sets"."purchase_date",
     "bricktracker_sets"."purchase_location",
+    "bricktracker_sets"."purchase_price",
     "rebrickable_sets"."set",
     "rebrickable_sets"."number",
     "rebrickable_sets"."version",